+++ main.m
@@ -1,1 +1,3 @@
-f_p1();
+function main()
+    f_p1();
+end